Refactor subkernel management to match context/agent management
