--- beaker_kernel/lib/subkernels/__init__.py
@@ -1,5 +1,0 @@
-from beaker_kernel.lib.autodiscovery import autodiscover
-from .base import BeakerSubkernel
-
-def autodiscover_subkernels():
-    return autodiscover("subkernels")
