+++ beaker_kernel/service/handlers.py
@@ -22,7 +22,7 @@ from tornado.web import StaticFileHandler, RedirectHandler, RequestHandler, HTTP
 from beaker_kernel.lib.autodiscovery import autodiscover
 from beaker_kernel.lib.app import BeakerApp
 from beaker_kernel.lib.context import BeakerContext
-from beaker_kernel.lib.subkernels.base import BeakerSubkernel
+from beaker_kernel.lib.subkernel import BeakerSubkernel
 from beaker_kernel.lib.agent_tasks import summarize
 from beaker_kernel.lib.config import config, locate_config, Config, Table, Choice, recursiveOptionalUpdate, reset_config
 from beaker_kernel.service import admin_utils
