+++ beaker_kernel/subkernels/rlang.py
@@ -4,7 +4,7 @@ import logging
 import re
 from typing import Any
 
-from .base import BeakerSubkernel
+from ..lib.subkernel import BeakerSubkernel
 
 logger = logging.getLogger(__name__)
